Week_3/29092025/jwt_fastapi/app/security.py
import os
from datetime import datetime, timedelta, timezone
from jose import jwt # type: ignore

SECRET_KEY = os.getenv("SECRET_KEY", "a_very_long_and_secure_secret_key_for_your_banking_app")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "60"))

def create_access_token(data: dict) -> str:
    to_encode = data.copy()
    to_encode.update({"exp": datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
