Week_3/29092025/jwt_fastapi/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic
requests
